Refactor display handling and improve UI responsiveness
- Updated display configuration to support new dimensions and panel orientation for the ST7735 display.
- Enhanced UI rendering logic to minimize unnecessary updates, improving performance and reducing flicker.
- Adjusted button mapping for better user interaction and refined encoder handling for smoother input.
- Implemented GPIO handling via lgpio for improved reliability and performance in hardware interactions.
- Removed obsolete documentation files to streamline project structure.

+++ firmware/src/upright/hal/button.py
@@ -52,7 +52,7 @@ def _map_pattern(state: _BtnState, pattern: str) -> str:
 
 def _loop(evt_bus: EventBus, stop: threading.Event) -> None:  # pragma: no cover
     states = [
-        _BtnState("a", PIN_BUTTON_A, map_short="long", map_long="long", map_verylong="verylong"),
+        _BtnState("a", PIN_BUTTON_A, map_short="single", map_long="long", map_verylong="verylong"),
         _BtnState("b", PIN_BUTTON_B, map_short="single", map_long="double", map_verylong="triple"),
     ]
     for st in states:
